Attempt to generalize this layout algorithm
