+++ modules/OrderedLayout/pom.xml
@@ -12,33 +12,31 @@
     <version>1.0.0</version>
     <packaging>nbm</packaging>
 
-    <name>EventGraphLayout</name>
+    <name>OrderedGraphLayout</name>
 
 
     <dependencies>
         <!-- Insert dependencies here -->
-
         <dependency>
             <groupId>org.gephi</groupId>
             <artifactId>layout-api</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>org.gephi</groupId>
-            <artifactId>ui-propertyeditor</artifactId>
+            <version>0.9.2</version>
         </dependency>
         <dependency>
             <groupId>org.gephi</groupId>
             <artifactId>graph-api</artifactId>
+            <version>0.9.2</version>
         </dependency>
         <dependency>
-            <groupId>org.netbeans.api</groupId>
-            <artifactId>org-openide-util-lookup</artifactId>
+            <groupId>org.gephi</groupId>
+            <artifactId>ui-propertyeditor</artifactId>
+            <version>0.9.2</version>
         </dependency>
         <dependency>
             <groupId>org.netbeans.api</groupId>
             <artifactId>org-openide-util</artifactId>
+            <version>RELEASE81</version>
         </dependency>
-
     </dependencies>
 
     <build>
