+++ pom.xml
@@ -63,6 +63,24 @@
             <classifier>linux</classifier>
             <type>tar.gz</type>
         </dependency>
+        <dependency>
+            <groupId>${project.groupId}</groupId>
+            <artifactId>layout-api</artifactId>
+            <version>0.9.2</version>
+            <type>nbm</type>
+        </dependency>
+        <dependency>
+            <groupId>${project.groupId}</groupId>
+            <artifactId>graph-api</artifactId>
+            <version>0.9.2</version>
+            <type>nbm</type>
+        </dependency>
+        <dependency>
+            <groupId>${project.groupId}</groupId>
+            <artifactId>project-api</artifactId>
+            <version>0.9.2</version>
+            <type>nbm</type>
+        </dependency>
     </dependencies>
     
     <build>
